# openapi.yaml
openapi: 3.1.0
info:
  title: GPT Drive Connector
  version: "1.0.0"
servers:
  - url: https://YOUR-RAILWAY-URL   # replace after deploy
paths:
  /health:
    get:
      summary: Health check
      operationId: health
      security: []
      responses:
        "200":
          description: ok
  /drive/files:
    get:
      summary: List files within the allowed folder
      operationId: listFiles
      parameters:
        - in: query
          name: folderId
          schema:
            type: string
          description: Optional, defaults to ALLOWED_FOLDER_ID
        - in: query
          name: pageSize
          schema:
            type: integer
            default: 25
      responses:
        "200":
          description: List of files
  /drive/search:
    get:
      summary: Search files by filename (within allowed folder)
      operationId: searchFiles
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
        - in: query
          name: pageSize
          schema:
            type: integer
            default: 25
      responses:
        "200":
          description: List of files
  /file/{id}/meta:
    get:
      summary: Get file metadata (only if file is in allowed folder)
      operationId: getFileMeta
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: File metadata
        "403":
          description: File outside allowed folder
  /file/{id}/text:
    get:
      summary: Extract text from a Google Doc (as plain text) or PDF
      operationId: extractText
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Extracted text JSON
        "403":
          description: File outside allowed folder
        "415":
          description: Unsupported mimeType
components:
  securitySchemes:
    googleOAuth:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://accounts.google.com/o/oauth2/auth
          tokenUrl: https://oauth2.googleapis.com/token
          scopes:
            https://www.googleapis.com/auth/drive.readonly: Read-only access to Google Drive files
security:
  - apiKey: []
  - googleOAuth: [https://www.googleapis.com/auth/drive.readonly]
components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: x-api-key
    googleOAuth:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://accounts.google.com/o/oauth2/auth
          tokenUrl: https://oauth2.googleapis.com/token
          scopes:
            https://www.googleapis.com/auth/drive.readonly: Read-only access to Google Drive files